nit picks and more refactoring
